+++ evalml/pipelines/components/estimators/regressors/varmax_regressor.py
@@ -34,6 +34,8 @@ class VARMAXRegressor(Estimator):
             solely be based off of the datetimes and target values. Defaults to True.
     """
 
+    _N_REPETITIONS = 400
+
     name = "VARMAX Regressor"
     hyperparameter_ranges = {
         "p": Integer(0, 10),
@@ -215,11 +217,51 @@ class VARMAXRegressor(Estimator):
             predictions (pd.Series): Not used for VARMAX regressor.
 
         Returns:
-            dict: Prediction intervals, keys are in the format {coverage}_lower or {coverage}_upper.
+            dict[dict]: A dict of prediction intervals, where the dict is in the format {series_id: {coverage}_lower or {coverage}_upper}.
         """
-        raise NotImplementedError(
-            "VARMAX does not have prediction intervals implemented yet.",
+        if coverage is None:
+            coverage = [0.95]
+
+        X, y = self._manage_woodwork(X, y)
+        use_exog = (
+            # If exogenous variables were used during training
+            self._component_obj._fitted_forecaster.model.exog is not None
+            and self.use_covariates
         )
+        if use_exog:
+            X = X.ww.select(exclude=["Datetime"])
+            X = convert_bool_to_double(X)
+        # Accesses the fitted statsmodels model within sktime
+        # nsimulations represents how many steps should be simulated
+        # repetitions represents the number of simulations that should be run (confusing, I know)
+        # anchor represents where the simulations should start from (forecasting is done from the "end")
+        y_pred = self._component_obj._fitted_forecaster.simulate(
+            nsimulations=X.shape[0],
+            repetitions=self._N_REPETITIONS,
+            anchor="end",
+            random_state=self.random_seed,
+            exog=X if use_exog else None,
+        )
+        prediction_interval_result = {}
+        # Access the target column names (i.e. the series_id values) that the VARMAX component obj was fitted on
+        for series in self._component_obj._fitted_forecaster.model.endog_names:
+            series_result = {}
+            series_preds = y_pred[[col for col in y_pred.columns if series in col]]
+            for conf_int in coverage:
+                prediction_interval_lower = series_preds.quantile(
+                    q=round((1 - conf_int) / 2, 3),
+                    axis="columns",
+                )
+                prediction_interval_upper = series_preds.quantile(
+                    q=round((1 + conf_int) / 2, 3),
+                    axis="columns",
+                )
+                prediction_interval_lower.index = X.index
+                prediction_interval_upper.index = X.index
+                series_result[f"{conf_int}_lower"] = prediction_interval_lower
+                series_result[f"{conf_int}_upper"] = prediction_interval_upper
+            prediction_interval_result[series] = series_result
+        return prediction_interval_result
 
     @property
     def feature_importance(self) -> np.ndarray: